Splitting the table into a separate component to make it reusable in the review screen - and also displaying it in the review screen.
