+++ libs/application/templates/parental-leave/src/fields/Review/review-groups/Employment.tsx
@@ -1,5 +1,6 @@
 import {
   DataValue,
+  Label,
   RadioValue,
   ReviewGroup,
   formatPhoneNumber,
@@ -9,6 +10,7 @@ import { ReviewGroupProps } from './props'
 import { useLocale } from '@island.is/localization'
 import { NO, YES, parentalLeaveFormMessages } from '../../..'
 import { useStatefulAnswers } from '../../../hooks/useStatefulAnswers'
+import { EmployersTable } from '../../components/EmployersTable'
 
 export const Employment = ({
   application,
@@ -21,7 +23,6 @@ export const Employment = ({
       isSelfEmployed,
       isReceivingUnemploymentBenefits,
       unemploymentBenefits,
-      // TODO: Populate data with employers
       employers,
     },
   ] = useStatefulAnswers(application)
@@ -37,45 +38,50 @@ export const Employment = ({
             label={formatMessage(parentalLeaveFormMessages.selfEmployed.title)}
             value={isSelfEmployed}
           />
-          {isSelfEmployed === NO && isReceivingUnemploymentBenefits === NO && (
-            <Box paddingTop={2}>
-              <Text variant="default">Wait wait wait!</Text>
-            </Box>
-          )}
-        </GridColumn>
-
-        <GridColumn span={['12/12', '12/12', '12/12', '5/12']}>
-          {isSelfEmployed === NO && isReceivingUnemploymentBenefits === NO && (
-            <Box paddingTop={2}>
-              <Text variant="default">Wait wait wait!</Text>
-            </Box>
-          )}
         </GridColumn>
       </GridRow>
       {isSelfEmployed === NO && ( // only show benefits in review if user had to answer that question
         <GridRow>
           <GridColumn span={['12/12', '12/12', '12/12', '5/12']}>
-            <RadioValue
-              label={formatMessage(
-                parentalLeaveFormMessages.employer
-                  .isReceivingUnemploymentBenefitsTitle,
-              )}
-              value={isReceivingUnemploymentBenefits}
-            />
+            <Box paddingTop={2} paddingBottom={2}>
+              <RadioValue
+                label={formatMessage(
+                  parentalLeaveFormMessages.employer
+                    .isReceivingUnemploymentBenefitsTitle,
+                )}
+                value={isReceivingUnemploymentBenefits}
+              />
+            </Box>
           </GridColumn>
 
           <GridColumn span={['12/12', '12/12', '12/12', '5/12']}>
             {isReceivingUnemploymentBenefits === YES && (
-              <DataValue
-                label={formatMessage(
-                  parentalLeaveFormMessages.employer.unemploymentBenefits,
-                )}
-                value={unemploymentBenefits}
-              />
+              <Box paddingTop={2} paddingBottom={2}>
+                <DataValue
+                  label={formatMessage(
+                    parentalLeaveFormMessages.employer.unemploymentBenefits,
+                  )}
+                  value={unemploymentBenefits}
+                />
+              </Box>
             )}
           </GridColumn>
         </GridRow>
       )}
+      <GridRow>
+        <GridColumn span={['12/12', '12/12', '12/12', '5/12']}>
+          <Label>
+            {formatMessage(parentalLeaveFormMessages.employer.title)}
+          </Label>
+        </GridColumn>
+        <GridColumn span={['12/12', '12/12', '12/12', '12/12']}>
+          {employers?.length > 0 && (
+            <Box paddingTop={2}>
+              <EmployersTable employers={employers} />
+            </Box>
+          )}
+        </GridColumn>
+      </GridRow>
     </ReviewGroup>
   )
 }
