+++ libs/application/templates/parental-leave/src/lib/dataSchema.ts
@@ -72,32 +72,32 @@ export const dataSchema = z.object({
   shareInformationWithOtherParent: z.enum([YES, NO]),
   useUnion: z.enum([YES, NO]),
   usePrivatePensionFund: z.enum([YES, NO]),
-  employers: z.array(
-    z.object({
-      name: z.object({
-        label: z.string(),
-        nationalId: z.string().refine((n) => kennitala.isCompany(n), {
-          params: errorMessages.employerNationalRegistryId,
-        }),
-      }),
-      phoneNumber: z
-        .string()
-        .refine(
-          (p) => {
-            const phoneNumber = parsePhoneNumberFromString(p, 'IS')
-            if (phoneNumber) return phoneNumber.isValid()
-            else return true
-          },
-          { params: errorMessages.phoneNumber },
-        )
-        .optional(),
-      email: z.string().email(),
-      ratio: z.string().refine((value) => {
-        const intValue = parseInt(value)
-        return intValue >= 0 && intValue <= 100
-      }),
-    }),
-  ),
+  // employers: z.array(
+  //   z.object({
+  //     name: z.object({
+  //       label: z.string(),
+  //       nationalId: z.string().refine((n) => kennitala.isCompany(n), {
+  //         params: errorMessages.employerNationalRegistryId,
+  //       }),
+  //     }),
+  //     phoneNumber: z
+  //       .string()
+  //       .refine(
+  //         (p) => {
+  //           const phoneNumber = parsePhoneNumberFromString(p, 'IS')
+  //           if (phoneNumber) return phoneNumber.isValid()
+  //           else return true
+  //         },
+  //         { params: errorMessages.phoneNumber },
+  //       )
+  //       .optional(),
+  //     email: z.string().email(),
+  //     ratio: z.string().refine((value) => {
+  //       const intValue = parseInt(value)
+  //       return intValue >= 0 && intValue <= 100
+  //     }),
+  //   }),
+  // ),
   isReceivingUnemploymentBenefits: z.enum([YES, NO]),
   unemploymentBenefits: z.string().min(1),
   requestRights: z.object({
